+++ mindbox/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.0.0
+
+* Upgrade iOS SDK dependency to v2.0.1.
+* Upgrade Android SDK dependency to v2.1.1. 
+* Add support of Huawei Push Kit.
+* Update onPushClickReceived method: add second argument with push payload.
+
 ## 1.0.2
 
 * Add iOS old format push notifications support.